streamlit-app/requirements.txt
pesto-cli
streamlit
jsonschema
rasterio
numpy